% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggscatterstats.R
\name{bf_corr_test}
\alias{bf_corr_test}
\title{Bayesian correlation test.}
\usage{
bf_corr_test(data, x, y, bf.prior = 0.707, caption = NULL,
  output = "caption")
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{A vector containing the explanatory variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{bf.prior}{A number between 0.5 and 2 (default \code{0.707}), the prior width
to use in calculating Bayes factors.}

\item{caption}{The text for the plot caption.}

\item{output}{Can either be \code{"caption"} (which will contain text for evidence
in favor of null)  or \code{"results"} (which will return the dataframe with
results).}
}
\description{
Bayesian correlation test.
}
\author{
Indrajeet Patil
}
\keyword{internal}
