% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{bf_oneway_anova}
\alias{bf_oneway_anova}
\title{Bayesian one-way analysis of variance.}
\usage{
bf_oneway_anova(data, x, y, bf.prior = 0.707, caption = NULL,
  output = "caption")
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{bf.prior}{A number between 0.5 and 2 (default \code{0.707}), the prior width
to use in calculating Bayes factors.}

\item{caption}{The text for the plot caption.}

\item{output}{Can either be \code{"caption"} (which will contain text for evidence
in favor of null)  or \code{"results"} (which will return the dataframe with
results).}
}
\description{
Bayesian one-way analysis of variance.
}
\seealso{
\code{\link{bf_contigency_tab}}, \code{\link{bf_corr_test}}
}
\author{
Indrajeet Patil
}
\keyword{internal}
