% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcatstats.R
\name{cat_counter}
\alias{cat_counter}
\title{Preparing dataframe with counts and percentages for categorical
variables.}
\usage{
cat_counter(data, main, condition = NULL, ...)
}
\arguments{
\item{data}{The data as a data frame (matrix or tables will not be accepted).}

\item{main}{The variable to use as the \strong{rows} in the
contingency table.}

\item{condition}{The variable to use as the \strong{columns} in the contingency
table.}

\item{...}{Additional grouping variables.}
}
\description{
Preparing dataframe with counts and percentages for categorical
variables.
}
\examples{
ggstatsplot:::cat_counter(data = ggplot2::mpg, "drv", cyl, "fl")
}
\author{
Indrajeet Patil
}
\keyword{internal}
