% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcatstats.R
\name{cat_summary_label_maker}
\alias{cat_summary_label_maker}
\title{Summary dataframe for categorical variables.}
\usage{
cat_summary_label_maker(data, label.col.name = "slice.label",
  label.content = "percentage", perc.k = 1)
}
\arguments{
\item{data}{A dataframe containing summaries for categorical variables.
Should contain columns named either \code{"perc"} or \code{"counts"} or both.}

\item{label.col.name}{Character that decides the column name containing
summary label. This can either be \code{"slice.label"} (default) or
\code{"data.label"}.}

\item{label.content}{Character decides what information needs to be displayed
on the label in each pie or bar slice. Possible options are \code{"percentage"}
(default), \code{"counts"}, \code{"both"}.}

\item{perc.k}{Numeric that decides number of decimal places for percentage
labels (Default: \code{0}).}
}
\description{
Creating a dataframe with an added column corresponding to
summary for categorical variables.
}
\examples{
\dontrun{
# creating a dataframe with counts and percentage
df <-
  mtcars \%>\%
  dplyr::group_by(.data = ., am, cyl) \%>\%
  dplyr::summarize(.data = ., counts = n()) \%>\%
  dplyr::mutate(.data = ., perc = (counts / sum(counts)) * 100) \%>\%
  dplyr::ungroup(x = .) \%>\%
  dplyr::arrange(.data = ., dplyr::desc(x = cyl)) \%>\%
  dplyr::filter(.data = ., counts != 0L)

# dataframe with label column
ggstatsplot:::cat_summary_label_maker(
  data = df,
  label.col.name = "slice.label",
  label.content = "both",
  perc.k = 1
)
}

}
\author{
Indrajeet Patil
}
\keyword{internal}
