% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_w_ci.R
\name{chi_ncp_ci}
\alias{chi_ncp_ci}
\title{Computing confidence intervals for chi-squared statistic.}
\usage{
chi_ncp_ci(chi.square, conf.level = 0.95, df, tol = 1e-09,
  jumping.prop = 0.1, ...)
}
\arguments{
\item{chi.square}{A value of chi-squared statistic.}

\item{conf.level}{Scalar between 0 and 1. If unspecified, the defaults return
\code{95\%} lower and upper confidence intervals (\code{0.95}).}

\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{tol}{Tolerance.}

\item{jumping.prop}{Jumping proportion.}

\item{...}{Additional arguments.}
}
\description{
Computing confidence intervals for chi-squared statistic.
}
\author{
Chuck Powell
}
\keyword{internal}
