% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntPartialEtaSq.R
\name{confIntPartialEtaSq}
\alias{confIntPartialEtaSq}
\alias{partialeta_sq_ci}
\title{Confidence intervals for Omega Squared}
\usage{
partialeta_sq_ci(lm_object, conf.level = 0.95)
}
\arguments{
\item{lm_object}{stats::lm linear model object}

\item{conf.level}{Level of confidence for the confidence interval}
}
\author{
Indrajeet Patil

#'
}
