% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_palette_sufficient}
\alias{.is_palette_sufficient}
\title{Check if palette has enough number of colors}
\usage{
.is_palette_sufficient(palette, min_length)
}
\description{
Aborts with an informative error if the number of factor levels exceeds the
number of colors available in the specified palette.
}
\examples{
ggstatsplot:::.is_palette_sufficient("ggthemes::gdoc", 6L)
try(ggstatsplot:::.is_palette_sufficient("ggthemes::gdoc", 30L))

}
\keyword{internal}
