% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcoefstats_graphics.R
\name{easystats_to_tidy_names}
\alias{easystats_to_tidy_names}
\title{Convert \code{easystats} package outputs to \code{tidymodels} conventions.}
\usage{
easystats_to_tidy_names(x)
}
\description{
Both \code{broom} package from \code{tidymodels} universe and \code{parameters}
package from \code{easystats} universe can provide model summaries for a large
number of model objects. This is a convenience function that converts
naming conventions adopted in \code{easystats} to the ones adopted in the
\code{broom} package.
}
\examples{
# example model object
mod <- stats::lm(formula = wt ~ am * cyl, data = mtcars)

# `tidy`-fied output
ggstatsplot:::easystats_to_tidy_names(parameters::model_parameters(mod))

}
\keyword{internal}
