% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_t_subtitles.R
\name{effect_t_parametric}
\alias{effect_t_parametric}
\title{Calculating Cohen or Hedges (between-/within- or one
sample designs).}
\usage{
effect_t_parametric(formula = NULL, data = NULL, mu = 0,
  paired = FALSE, hedges.correction = TRUE, conf.level = 0.95,
  noncentral = TRUE)
}
\arguments{
\item{formula}{this function only accepts the variables in
\code{formula} format e.g. \code{sleep_rem ~ vore} or \code{~ vore}.}

\item{data}{a data frame or tibble containing the data}

\item{mu}{if conducting a single sample test against a mean (Default: \code{0}).}

\item{paired}{Logical indicating whether to treat the data as a oaired sample
e.g. \code{post ~ pre}.}

\item{hedges.correction}{Logical indicating whether to apply Hedges correction,
Hedge's \emph{g} (Default: \code{TRUE}).}

\item{conf.level}{A scalar value between 0 and 1. If unspecified, the
default is to return \code{95\%} lower and upper confidence intervals (\code{0.95}).}

\item{noncentral}{Logical indicating whether to use non-central
\emph{t}-distributions for computing the confidence intervals (Default: \code{TRUE}).}
}
\description{
Calculating Cohen or Hedges (between-/within- or one
sample designs).
}
\examples{

# creating a smaller dataset
msleep_short <- dplyr::filter(
  .data = ggplot2::msleep,
  vore \%in\% c("carni", "herbi")
)

# with defaults
effect_t_parametric(
  formula = sleep_rem ~ vore,
  data = msleep_short,
)

# changing defaults
effect_t_parametric(
  formula = sleep_rem ~ vore,
  data = msleep_short,
  mu = 1, # ignored in this case
  paired = FALSE,
  hedges.correction = TRUE,
  conf.level = .99,
  noncentral = FALSE
)
}
\author{
Chuck Powell
}
