% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{effsize_ci_message}
\alias{effsize_ci_message}
\title{Message about bootstrapped confidence intervals for effect sizes.}
\usage{
effsize_ci_message(nboot = 100, conf.level = 0.95)
}
\arguments{
\item{nboot}{Number of bootstrap samples for computing confidence interval
for the effect size (Default: \code{100}).}

\item{conf.level}{Scalar between 0 and 1. If unspecified, the defaults return
\code{95\%} lower and upper confidence intervals (\code{0.95}).}
}
\description{
Message about bootstrapped confidence intervals for effect sizes.
}
\seealso{
Other helper_messages: \code{\link{bartlett_message}},
  \code{\link{ggcorrmat_matrix_message}},
  \code{\link{grouped_message}},
  \code{\link{normality_message}},
  \code{\link{pairwise_p}}, \code{\link{palette_message}}
}
\author{
Indrajeet Patil
}
\concept{helper_messages}
\keyword{internal}
