% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_functions.R
\name{effsize_type_switch}
\alias{effsize_type_switch}
\title{Switch function to determine which effect size is to computed.}
\usage{
effsize_type_switch(effsize.type)
}
\arguments{
\item{effsize.type}{Character describing the needed effect size.}
}
\description{
Takes in all allowed characters describing the needed effect
size (e.g., \code{"d"}, \code{"partial_eta"}, etc.) and converts it into standard
terms (\code{"biased"} or \code{"unbiased"}) to reduce the complexity of conditional
statements.
}
\author{
Indrajeet Patil
}
\keyword{internal}
