% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grouped_stats.R
\name{extract_grouped_stats}
\alias{extract_grouped_stats}
\title{Get stats from grouped plots}
\usage{
extract_grouped_stats(p = NULL)
}
\arguments{
\item{p}{A grouped plot made with grouped_ggwithinstats()}
}
\value{
A tibble summarising stats for a grouped plot made with ggstastplot
}
\description{
Get stats from grouped plots
}
