% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-stats.R
\name{extract_stats}
\alias{extract_stats}
\alias{extract_subtitle}
\alias{extract_caption}
\title{Extracting data frames or expressions from \code{{ggstatsplot}} plots}
\usage{
extract_stats(p)

extract_subtitle(p)

extract_caption(p)
}
\arguments{
\item{p}{A plot from \code{{ggstatsplot}} package}
}
\value{
For \code{extract_stats()}, a list with the following elements, each either a
data frame or \code{NULL} if not relevant for the given plot:
\itemize{
\item \code{subtitle_data}: results of the test displayed in the subtitle
\item \code{caption_data}: results of the test displayed in the caption (typically,
the Bayesian test)
\item \code{pairwise_comparisons_data}: results of pairwise comparisons
\item \code{descriptive_data}: counts and percentages (\code{ggpiestats()} and
\code{ggbarstats()})
\item \code{one_sample_data}: proportion tests for each level of \code{y}
(\code{ggpiestats()} and \code{ggbarstats()})
\item \code{tidy_data}: tidy model parameters (\code{ggcoefstats()})
\item \code{glance_data}: model performance indices (\code{ggcoefstats()})
}

For \code{extract_subtitle()} and \code{extract_caption()}, the expression displayed
in the subtitle or caption, respectively (or \code{NULL}).

For plots from \verb{grouped_*} functions, a list with one such element per plot.
}
\description{
Extracting data frames or expressions from \code{{ggstatsplot}} plots
}
\details{
These are convenience functions to extract data frames or expressions with
statistical details that are used to create expressions displayed in
\code{{ggstatsplot}} plots as subtitle, caption, etc. Note that all of this
analysis is carried out by the \code{{statsExpressions}}
\href{https://www.indrapatil.com/statsExpressions/}{package}. And so if you
are using these functions only to extract data frames, you are better off
using that package.

The only exception is the \code{ggcorrmat()} function, for which all elements
are \code{NULL}. But, if a data frame is what you want, you shouldn't be using
\code{ggcorrmat()} anyway. You can use \code{\link[correlation:correlation]{correlation::correlation()}} function
which provides tidy data frames by default.
}
\examples{
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true")) withAutoprint(\{ # examplesIf}
set.seed(123)

# non-grouped plot
p1 <- ggbetweenstats(mtcars, cyl, mpg)

# grouped plot
p2 <- grouped_ggbarstats(Titanic_full, Survived, Sex, grouping.var = Age)

# extracting expressions -----------------------------

extract_subtitle(p1)
extract_caption(p1)

extract_subtitle(p2)
extract_caption(p2)

# extracting data frames -----------------------------

extract_stats(p1)

extract_stats(p2)
\dontshow{\}) # examplesIf}
}
