% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pairwise_comparison.R
\name{games_howell}
\alias{games_howell}
\title{Games-Howell post-hoc test}
\usage{
games_howell(data, x, y)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}
}
\description{
This function produces results from Games-Howell post-hoc tests
for Welch's one-way analysis of variance (ANOVA) (\code{stats::oneway.test()}).
}
\note{
This is based on the implementation of Games-Howell test by Aaron
Schlegel (https://rpubs.com/aaronsc32)  and published on RPubs
(https://rpubs.com/aaronsc32/games-howell-test).
}
\keyword{internal}
