% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_functions.R
\name{ggbetweenstats_caption_switch}
\alias{ggbetweenstats_caption_switch}
\alias{ggwithinstats_caption_switch}
\title{Switch function to use helper function to create caption for the
\code{ggbetweenstats} plot.}
\usage{
ggbetweenstats_caption_switch(test, ...)

ggwithinstats_caption_switch(test, ...)
}
\arguments{
\item{test}{Decides which test to run (can be either \code{"t"} or
\code{"anova"}).}

\item{...}{Arguments passed to respective caption helper functions.}
}
\description{
Switch function to use helper function to create caption for the
\code{ggbetweenstats} plot.
}
\keyword{internal}
