% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_functions.R
\name{ggbetweenstats_subtitle_switch}
\alias{ggbetweenstats_subtitle_switch}
\alias{ggwithinstats_subtitle_switch}
\title{Switch function to use helper function to create subtitle for the
\code{ggbetweenstats} plot.}
\usage{
ggbetweenstats_subtitle_switch(type, test, ...)

ggwithinstats_subtitle_switch(type, test, ...)
}
\arguments{
\item{type}{Type of statistic expected (\code{"parametric"} or \code{"nonparametric"}
or \code{"robust"} or \code{"bayes"}).Corresponding abbreviations are also accepted:
\code{"p"} (for parametric), \code{"np"} (nonparametric), \code{"r"} (robust), or
\code{"bf"}resp.}

\item{test}{Decides which test to run (can be either \code{"t"} or
\code{"anova"}).}

\item{...}{Arguments passed to respective subtitle helper functions.}
}
\description{
Switch function to use helper function to create subtitle for the
\code{ggbetweenstats} plot.
}
\keyword{internal}
