% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{ggcorrmat_matrix_message}
\alias{ggcorrmat_matrix_message}
\title{Message to display when adjusted p-values are displayed in correlation
matrix.}
\usage{
ggcorrmat_matrix_message()
}
\description{
Message to display when adjusted p-values are displayed in correlation
matrix.
}
\keyword{internal}
