% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatterstat.R
\name{ggscatterstat}
\alias{ggscatterstat}
\title{scatterplot with ggMarginals}
\arguments{
\item{df}{data frame from which variables specified are preferentially to be taken}

\item{x}{a vector containing the explanatory variable}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{y}{the response - a vector of length the number of rows of `x`}

\item{marginal}{decides whether `ggExtra::ggMarginal()` plots will be displayed; the default is `TRUE`}

\item{marginaltype}{type of marginal distribution to be plotted on the axes (NULL, "histogram", "boxplot", "density")}

\item{xfill}{color fill for x axis distibution}

\item{yfill}{color fill for y axis distribution}

\item{test}{statistical test to be run and displayed as subtitle ("pearson", "spearman", "robust")}

\item{intercept}{decides whether "mean" or "median" or no intercept lines (`NULL`) are to be plotted}

\item{title}{title for the plot}

\item{caption}{caption for the plot}
}
\author{
Indrajeet Patil
}
