% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{ggsignif_position_calculator}
\alias{ggsignif_position_calculator}
\title{Calculating \code{y} coordinates for the \code{ggsignif} comparison bars.}
\usage{
ggsignif_position_calculator(x, y)
}
\arguments{
\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}
}
\description{
Calculating \code{y} coordinates for the \code{ggsignif} comparison bars.
}
\keyword{internal}
