% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstatsplot.R
\docType{package}
\name{ggstatsplot}
\alias{ggstatsplot}
\alias{ggstatsplot-package}
\title{\code{ggstatsplot}}
\description{
Collection of functions to enhance \code{ggplot2} plots with results from
statistical tests.
}
