% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{grouped_message}
\alias{grouped_message}
\title{grouped_message}
\usage{
grouped_message()
}
\description{
A note to the user about the class of the output object.
}
\seealso{
\code{\link{grouped_ggbetweenstats}},
\code{\link{grouped_gghistostats}}, \code{\link{grouped_ggscatterstats}},
\code{\link{grouped_ggpiestats}}, \code{\link{grouped_ggcorrmat}}

Other helper_messages: \code{\link{bartlett_message}},
  \code{\link{ggcorrmat_matrix_message}},
  \code{\link{normality_message}},
  \code{\link{pairwise_p}}, \code{\link{palette_message}}
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
\concept{helper_messages}
\keyword{internal}
