% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{long_to_wide_converter}
\alias{long_to_wide_converter}
\title{Converts long-format dataframe to wide-format dataframe}
\usage{
long_to_wide_converter(data, x, y)
}
\description{
Converts long-format dataframe to wide-format dataframe
}
\author{
Indrajeet Patil
}
\keyword{internal}
