% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{mean_labeller}
\alias{mean_labeller}
\title{Create a dataframe with mean per group and a formatted label for
display in \code{ggbetweenstats} plot.}
\usage{
mean_labeller(data, x, y, mean.ci = FALSE, k = 3)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{mean.ci}{Logical that decides whether 95% confidence interval for mean
is to be displayed (Default: \code{FALSE}).}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 3}).}
}
\description{
Create a dataframe with mean per group and a formatted label for
display in \code{ggbetweenstats} plot.
}
\keyword{internal}
