% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{normality_message}
\alias{normality_message}
\title{Display normality test result as a message.}
\usage{
normality_message(x, lab, k = 3, output = "message")
}
\arguments{
\item{x}{A numeric vector}

\item{lab}{A character describing label for the variable.}

\item{k}{Number of decimal places expected for results (Default: \code{3}).}

\item{output}{What output is desired: \code{"message"} (default) or \code{"stats"} objects.}
}
\description{
A note to the user about the validity of assumptions for the
default linear model.
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::normality_message}. Note that it is \code{:::} and not \code{::}.
}
\keyword{internal}
