% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcorrmat.R
\name{numdf_n_summary}
\alias{numdf_n_summary}
\title{Compute minimum, maximum, and median for a dataframe with numbers.}
\usage{
numdf_n_summary(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A tibble with minimum, median, and maximum values.
}
\description{
Compute minimum, maximum, and median for a dataframe with numbers.
}
\keyword{internal}
