% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcorrmat.R
\name{numdf_summary}
\alias{numdf_summary}
\title{Compute minimum, maximum, and median across all numeric columns in a
dataframe.}
\usage{
numdf_summary(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A tibble with minimum, median, and maximum values.
}
\description{
Compute minimum, maximum, and median across all numeric columns in a
dataframe.
}
\examples{
\donttest{
ggstatsplot:::numdf_summary(ggplot2::msleep)
}

}
\author{
Indrajeet Patil
}
\keyword{internal}
