% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats_graphics.R
\name{outlier_df}
\alias{outlier_df}
\title{Adding a column to dataframe describing outlier status}
\usage{
outlier_df(data, x, y, outlier.label, outlier.coef = 1.5, ...)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{outlier.label}{Label to put on the outliers that have been tagged. This
\strong{can't} be the same as \code{x} argument.}

\item{outlier.coef}{Coefficient for outlier detection using Tukey's method.
With Tukey's method, outliers are below (1st Quartile) or above (3rd
Quartile) \code{outlier.coef} times the Inter-Quartile Range (IQR) (Default:
\code{1.5}).}

\item{...}{Additional arguments.}
}
\description{
Adding a column to dataframe describing outlier status
}
\examples{
# adding column for outlier and a label for that outlier
ggstatsplot:::outlier_df(
  data = morley,
  x = Expt,
  y = Speed,
  outlier.label = Run,
  outlier.coef = 2
) \%>\%
  dplyr::arrange(outlier)
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
\keyword{internal}
