% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_functions.R
\name{p.adjust.method.description}
\alias{p.adjust.method.description}
\title{Preparing text to describe which \emph{p}-value adjustment method was used.}
\usage{
p.adjust.method.description(p.adjust.method)
}
\arguments{
\item{p.adjust.method}{Adjustment method for \emph{p}-values for multiple
comparisons. Possible methods are: \code{"holm"} (default), \code{"hochberg"},
\code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"}.}
}
\value{
Standardized text description for what method was used.
}
\description{
Preparing text to describe which \emph{p}-value adjustment method was used.
}
\keyword{internal}
