% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{pairwise_p_caption}
\alias{pairwise_p_caption}
\title{Preparing caption in case pairwise comparisons are displayed.}
\usage{
pairwise_p_caption(type, var.equal, paired, p.adjust.method,
  caption = NULL)
}
\arguments{
\item{type}{Type of statistic expected (\code{"parametric"} or \code{"nonparametric"}
or \code{"robust"} or \code{"bayes"}).Corresponding abbreviations are also accepted:
\code{"p"} (for parametric), \code{"np"} (nonparametric), \code{"r"} (robust), or
\code{"bf"}resp.}

\item{var.equal}{a logical variable indicating whether to treat the
    two variances as being equal. If \code{TRUE} then the pooled
    variance is used to estimate the variance otherwise the Welch
    (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{paired}{a logical indicating whether you want a paired
    t-test.}

\item{p.adjust.method}{Adjustment method for \emph{p}-values for multiple
comparisons. Possible methods are: \code{"holm"} (default), \code{"hochberg"},
\code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"}.}

\item{caption}{The text for the plot caption.}
}
\description{
Preparing caption in case pairwise comparisons are displayed.
}
\keyword{internal}
