% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcoefstats_graphics.R
\name{parameters_tidy}
\alias{parameters_tidy}
\title{Tidier for \code{parameters} package objects}
\usage{
parameters_tidy(x, ...)
}
\arguments{
\item{...}{Arguments passed to or from other methods.}
}
\description{
Tidier for \code{parameters} package objects
}
\examples{
\donttest{
# setup
library(lme4)
library(parameters)
set.seed(123)

# model
mm0 <-
  lme4::lmer(
    formula = scale(Reaction) ~ scale(Days) + (1 | Subject),
    data = sleepstudy
  )

# model parameters
ggstatsplot:::parameters_tidy(mm0)
}

}
\keyword{internal}
