% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialeta_sq_ci.R
\name{partialeta_sq_ci}
\alias{partialeta_sq_ci}
\title{Confidence intervals for Partial Eta Squared}
\usage{
partialeta_sq_ci(lm_object, conf.level = 0.95)
}
\arguments{
\item{lm_object}{stats::lm linear model object}

\item{conf.level}{Level of confidence for the confidence interval}
}
\author{
Indrajeet Patil
}
