% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{proptest_message}
\alias{proptest_message}
\title{Message about results from a single-sample proportion test.}
\usage{
proptest_message(main, condition)
}
\arguments{
\item{main}{The variable to use as the \strong{rows} in the
contingency table.}

\item{condition}{The variable to use as the \strong{columns} in the contingency
table.}
}
\description{
Message about results from a single-sample proportion test.
}
\examples{
\dontrun{
ggstatsplot:::proptest_message(main = am, condition = cyl)
}

}
\author{
Indrajeet Patil
}
\keyword{internal}
