% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{signif_column}
\alias{signif_column}
\title{Creating a new character type column with significance labels}
\usage{
signif_column(data = NULL, p)
}
\arguments{
\item{data}{Data frame from which variables specified are preferentially to
be taken.}

\item{p}{The column containing p-values.}
}
\value{
Returns the originally entered object (either a vector or a
dataframe) in tibble format with an additional column corresponding to
statistical significance.
}
\description{
This function will add a new column to a dataframe containing
p-values
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::signif_column}. Note that it is \code{:::} and not \code{::}.
}
\author{
Indrajeet Patil
}
\keyword{internal}
