% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal.R
\name{specify_decimal}
\alias{specify_decimal}
\title{custom function for getting specified number of decimal places in results}
\usage{
specify_decimal(x, k = NULL)
}
\arguments{
\item{x}{a numeric value}

\item{k}{the number of digits after decimal point (should be an integer)}
}
\value{
formatted values for results from statistical tests
}
