% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_anova_subtitles.R
\name{subtitle_friedman_nonparametric}
\alias{subtitle_friedman_nonparametric}
\title{Making text subtitle for the Friedman Rank Sum Test (nonparametric
ANOVA) (within-subjects designs).}
\usage{
subtitle_friedman_nonparametric(data, x, y, conf.level = 0.95, k = 2,
  messages = TRUE, ...)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{conf.level}{Scalar between 0 and 1. If unspecified, the defaults return
\code{95\%} lower and upper confidence intervals (\code{0.95}).}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2}).}

\item{messages}{Decides whether messages references, notes, and warnings are
to be displayed (Default: \code{TRUE}).}

\item{...}{Additional arguments (ignored).}
}
\description{
Making text subtitle for the Friedman Rank Sum Test (nonparametric
ANOVA) (within-subjects designs).
}
\examples{
# setup
set.seed(123)
library(ggstatsplot)
library(jmv)
data("bugs", package = "jmv")

# converting to long format
data_bugs <- bugs \%>\%
  tibble::as_tibble(.) \%>\%
  tidyr::gather(., key, value, LDLF:HDHF)

# creating the subtitle
ggstatsplot::subtitle_friedman_nonparametric(
  data = data_bugs,
  x = key,
  y = value,
  k = 2
)
}
\author{
Indrajeet Patil
}
