% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_subtitles.R
\name{subtitle_ggbetween_kw_nonparametric}
\alias{subtitle_ggbetween_kw_nonparametric}
\title{Making text subtitle for the Kruskal-Wallis test (nonparametric ANOVA)
(between-subjects designs).}
\usage{
subtitle_ggbetween_kw_nonparametric(data, x, y, messages = TRUE, k = 3,
  ...)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{messages}{Decides whether messages references, notes, and warnings are
to be displayed (Default: \code{TRUE}).}

\item{k}{Number of digits after decimal point (should be an integer) (Default: \code{k = 3}).}

\item{...}{Additional arguments (ignored).}
}
\description{
Making text subtitle for the Kruskal-Wallis test (nonparametric ANOVA)
(between-subjects designs).
}
\examples{
subtitle_ggbetween_kw_nonparametric(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem
)
}
\author{
Indrajeet Patil
}
