% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_subtitles.R
\name{subtitle_ggbetween_mann_nonparametric}
\alias{subtitle_ggbetween_mann_nonparametric}
\title{Making text subtitle for the Mann-Whitney U-test
(between-/within-subjects designs).}
\usage{
subtitle_ggbetween_mann_nonparametric(data, x, y, paired = FALSE,
  k = 3)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{paired}{a logical indicating whether you want a paired
    t-test.}

\item{k}{Number of digits after decimal point (should be an integer) (Default: \code{k = 3}).}
}
\description{
Making text subtitle for the Mann-Whitney U-test
(between-/within-subjects designs).
}
\author{
Indrajeet Patil
}
\keyword{internal}
