% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_subtitles.R
\name{subtitle_ggbetween_t_bayes}
\alias{subtitle_ggbetween_t_bayes}
\title{Making text subtitle for the between-subject bayesian t-test.}
\usage{
subtitle_ggbetween_t_bayes(data, x, y, bf.prior = 0.707, k = 3)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{bf.prior}{A number between 0.5 and 2 (default \code{0.707}), the prior width
to use in calculating Bayes factors.}

\item{k}{Number of digits after decimal point (should be an integer) (Default: \code{k = 3}).}
}
\description{
Making text subtitle for the between-subject bayesian t-test.
}
\author{
Indrajeet Patil
}
\keyword{internal}
