% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetween_subtitles.R
\name{subtitle_t_robust}
\alias{subtitle_t_robust}
\title{Making text subtitle for the robust t-test
(between- and within-subjects designs).}
\usage{
subtitle_t_robust(data, x, y, tr = 0.1, paired = FALSE, nboot = 100,
  conf.level = 0.95, conf.type = "norm", k = 3, messages = TRUE,
  ...)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{x}{The grouping variable.}

\item{y}{The response - a vector of length the number of rows of \code{x}.}

\item{tr}{Trim level for the mean when carrying out \code{robust} tests. If you
get error stating "Standard error cannot be computed because of Winsorized
variance of 0 (e.g., due to ties). Try to decrease the trimming level.",
try to play around with the value of \code{tr}, which is by default set to
\code{0.1}. Lowering the value might help.}

\item{paired}{a logical indicating whether you want a paired
    t-test.}

\item{nboot}{Number of bootstrap samples for computing effect size (Default:
\code{100}).}

\item{conf.level}{A scalar value between 0 and 1. If unspecified, the
default is to return \code{95\%} lower and upper confidence intervals (\code{0.95}).}

\item{conf.type}{A vector of character strings representing the type of
intervals required. The value should be any subset of the values \code{"norm"},
\code{"basic"}, \code{"perc"}, \code{"bca"}. For more, see \code{?boot::boot.ci}.}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 3}).}

\item{messages}{Decides whether messages references, notes, and warnings are
to be displayed (Default: \code{TRUE}).}

\item{...}{Additional arguments (ignored).}
}
\description{
Making text subtitle for the robust t-test
(between- and within-subjects designs).
}
\examples{

# with defaults
subtitle_t_robust(
  data = sleep,
  x = group,
  y = extra
)

# changing defaults
subtitle_t_robust(
  data = ToothGrowth,
  x = supp,
  y = len,
  nboot = 10,
  k = 1,
  tr = 0.2
)

# within-subjects design
ggstatsplot::subtitle_t_robust(
  data = dplyr::filter(
    ggstatsplot::intent_morality,
    condition \%in\% c("accidental", "attempted"),
    harm == "Poisoning"
  ),
  x = condition,
  y = rating,
  paired = TRUE,
  nboot = 25
)
}
\author{
Indrajeet Patil
}
