% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitle_maker_templates.R
\name{subtitle_template_1}
\alias{subtitle_template_1}
\title{Template for subtitles with statistical details for tests with a
single parameter (e.g., t, chi-squared, etc.)}
\usage{
subtitle_template_1(stat.title = NULL, statistic.text, statistic,
  parameter, p.value, effsize.text, effsize.estimate, effsize.LL,
  effsize.UL, n, conf.level = 0.95, k = 3L, k.parameter = 0L)
}
\arguments{
\item{stat.title}{A character describing the test being run, which will be
added as a prefix in the subtitle. The default is \code{NULL}. An example of a
\code{stat.title} argument will be something like \code{"Student's t-test: "}.}

\item{statistic.text}{A character specifiying the relevant test statistic.
For example, for tests with \emph{t}-statistic, \code{statistic.text = "t"}. If you
want to use plotmath, you will have to quote the argument (e.g.,
\code{quote(italic("t"))}).}

\item{statistic}{The numeric value of a statistic.}

\item{parameter}{The numeric value of a parameter being modeled (often
degrees of freedom for the test).}

\item{p.value}{The two-sided p-value associated with the observed statistic.}

\item{effsize.text}{A character specifiying the relevant effect size.
For example, for Cohen's \emph{d} statistic, \code{effsize.text = "d"}. If you
want to use plotmath, you will have to quote the argument (e.g.,
\code{quote(italic("d"))}).}

\item{effsize.estimate}{The estimated value of the effect size.}

\item{effsize.LL}{The estimated value of the lower bound of effect size.}

\item{effsize.UL}{The estimated value of the upper bound of effect size.}

\item{n}{An integer specifying the sample size used for the test.}

\item{k}{Number of decimal places to display (default: \code{3}).}

\item{k.parameter}{Number of decimal places to display for the parameter
(default: \code{0}).}
}
\description{
Template for subtitles with statistical details for tests with a
single parameter (e.g., t, chi-squared, etc.)
}
\examples{
set.seed(123)

# subtitle for *t*-statistic with Cohen's *d* as effect size
ggstatsplot:::subtitle_template_1(
  statistic.text = quote(italic("t")),
  statistic = 5.494,
  parameter = 29.234,
  p.value = 0.00001,
  effsize.text = quote(italic("d")),
  effsize.estimate = -1.980,
  effsize.LL = -2.873,
  effsize.UL = -1.088,
  n = 32L,
  conf.level = 0.95,
  k = 3L,
  k.parameter = 3L
)
}
\author{
Indrajeet Patil
}
\keyword{internal}
