% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_theme_ggstatsplot.R
\name{theme_corrmat}
\alias{theme_corrmat}
\title{Default theme used for correlation matrix}
\usage{
theme_corrmat()
}
\value{
A \code{ggplot2} object with the \code{theme_ggstatsplot} overlaid.
}
\description{
Default theme used for correlation matrix
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
