% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ggstatsplot.R
\name{theme_ggstatsplot}
\alias{theme_ggstatsplot}
\title{Default theme used in \code{{ggstatsplot}}}
\usage{
theme_ggstatsplot()
}
\value{
A \code{{ggplot2}} theme object.
}
\description{
Common theme used across all plots generated in \code{{ggstatsplot}} and \emph{assumed}
by the author to be aesthetically pleasing to the user. The theme is a
wrapper around \code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}}.

All \code{{ggstatsplot}} functions have a \code{ggtheme} parameter that lets you choose
a different theme.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_ggstatsplot()
}
