% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_graphics.R
\name{theme_pie}
\alias{theme_pie}
\title{Default theme used for pie chart}
\usage{
theme_pie()
}
\value{
A \code{ggplot2} object with the \code{theme_mprl} theme.
}
\description{
Default theme used for pie chart
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::theme_pie}. Note that it is \code{:::} and not \code{::}.
}
\author{
Indrajeet Patil
#'
}
\keyword{internal}
