% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_labeller.R
\name{tz_labeller}
\alias{tz_labeller}
\title{Create text labels}
\usage{
tz_labeller(tidy_df, glance_df, statistic, k)
}
\arguments{
\item{tidy_df}{Tidy dataframe from \code{broom::tidy}.}

\item{glance_df}{Glance model summary dataframe from \code{broom::glance}.}

\item{statistic}{Which statistic is to be displayed (either \code{"t"} or \code{"z"}).}

\item{k}{Number of decimal places expected for results displayed in labels.}
}
\description{
Create text labels
}
\keyword{internal}
