% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stats.R
\name{untable}
\alias{untable}
\title{Untable a dataset}
\usage{
untable(data, counts)
}
\arguments{
\item{data}{A data.frame to untable.}

\item{counts}{A column containing counts.}
}
\description{
Given a tabulated dataset, this will untabulate it by repeating
each row by the number of times it was repeated.
}
\examples{

# have a look at the Titanic_full dataset first
Titanic_full <- untable(data = as.data.frame(Titanic), counts = Freq)
dplyr::glimpse(Titanic_full)
}
\seealso{
Other helper_stats: \code{\link{check_outlier}},
  \code{\link{signif_column}}
}
\concept{helper_stats}
