% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{bartlett_message}
\alias{bartlett_message}
\title{Display homogeneity of variance test as a message}
\usage{
bartlett_message(data, lab, k = 3, output = "message")
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken.}

\item{lab}{A character describing label for the variable.}

\item{k}{Number of decimal places expected for results (Default: \code{3}).}

\item{output}{What output is desired: \code{"message"} (default) or \code{"stats"} objects.}
}
\description{
A note to the user about the validity of assumptions for the
default linear model.
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::bartlett_message}. Note that it is \code{:::} and not \code{::}.
}
\keyword{internal}
