% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bf_tests.R
\name{bf_extractor}
\alias{bf_extractor}
\title{Convenience function to extract bayes factors from test}
\usage{
bf_extractor(bf.object)
}
\arguments{
\item{bf.object}{An object from \code{BayesFactor} package test results.}
}
\description{
Convenience function to extract bayes factors from test
}
\examples{
bf_extractor(
  BayesFactor::correlationBF(
    x = iris$Sepal.Length,
    y = iris$Sepal.Width
  )
)
}
