% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bf_tests.R
\name{bf_extractor}
\alias{bf_extractor}
\title{Extract Bayes Factors from \code{BayesFactor} model object.}
\usage{
bf_extractor(bf.object, ...)
}
\arguments{
\item{bf.object}{An object from \code{BayesFactor} package test results.}

\item{...}{Currently ignored.}
}
\description{
Extract Bayes Factors from \code{BayesFactor} model object.
}
\examples{
# getting only bayes factors
ggstatsplot::bf_extractor(
  BayesFactor::anovaBF(
    formula = Sepal.Length ~ Species,
    data = iris,
    progress = FALSE
  )
)
}
