% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcatstats_graphics.R
\name{cat_counter}
\alias{cat_counter}
\title{Counts and percentages across grouping variables.}
\usage{
cat_counter(data, x, y = NULL, ...)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The variable to use as the \strong{rows} in the contingency table.}

\item{y}{The variable to use as the \strong{columns} in the contingency
table. Default is \code{NULL}. If \code{NULL}, one-sample proportion test (a goodness
of fit test) will be run for the \code{x} variable. Otherwise an appropriate
association test will be run. This argument can not be \code{NULL} for
\code{ggbarstats} function.}

\item{...}{Currently ignored.}
}
\description{
Counts and percentages across grouping variables.
}
\examples{
ggstatsplot:::cat_counter(data = ggplot2::mpg, drv, cyl)
}
\keyword{internal}
