% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcatstats.R
\name{cat_counter}
\alias{cat_counter}
\title{Preparing dataframe with counts and percentages for categorical
variables.}
\usage{
cat_counter(data, main, condition = NULL, ...)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{main}{The variable to use as the \strong{rows} in the contingency table.}

\item{condition}{The variable to use as the \strong{columns} in the contingency
table. Default is \code{NULL}. If \code{NULL}, one-sample proportion test (a goodness
of fit test) will be run for the \code{main} variable. Otherwise an appropriate
association test will be run.}

\item{...}{Additional grouping variables.}
}
\description{
Preparing dataframe with counts and percentages for categorical
variables.
}
\examples{
ggstatsplot:::cat_counter(data = ggplot2::mpg, "drv", cyl, "fl")
}
\author{
Indrajeet Patil
}
\keyword{internal}
