% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_palette_sufficient}
\alias{.is_palette_sufficient}
\title{Check if palette has enough number of colors}
\usage{
.is_palette_sufficient(package, palette, min_length)
}
\description{
Informs the user about not using the default color palette
when the number of factor levels is greater than 8, the maximum number of
colors allowed by \code{"Dark2"} palette from the \code{{RColorBrewer}} package.
}
\examples{
.is_palette_sufficient("RColorBrewer", "Dark2", 6L)
.is_palette_sufficient("RColorBrewer", "Dark2", 12L)

}
\keyword{internal}
