% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grouped_stats.R
\name{extract_grouped_stats}
\alias{extract_grouped_stats}
\title{Get stats from grouped plots}
\usage{
extract_grouped_stats(p = NULL)
}
\arguments{
\item{p}{A grouped ggstatsplot (made with any of the grouped_ functions)}
}
\value{
A tibble summarizing stats for a grouped plot made with ggstastplot
}
\description{
Get stats from grouped plots
}
\examples{
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set.seed(123)

# grouped plot
p <- grouped_ggbarstats(Titanic_full, Survived, Sex, grouping.var = Age)

extract_grouped_stats(p)
\dontshow{\}) # examplesIf}
}
