% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{extract_stats}
\alias{extract_stats}
\title{Extracting data frames from \code{{ggstatsplot}} plots}
\usage{
extract_stats(p, ...)
}
\arguments{
\item{p}{A plot from \code{{ggstatsplot}} package}

\item{...}{Ignored}
}
\value{
A list of tibbles containing summaries of various statistical analyses.
The exact details included will depend on the function.
}
\description{
Extracting data frames from \code{{ggstatsplot}} plots
}
\details{
This is a convenience function to extract data frames with statistical details
that are used to create expressions displayed in \code{{ggstatsplot}} plots as
subtitle, caption, etc. Note that all of this analysis is carried out by
the \code{{statsExpressions}} \href{https://indrajeetpatil.github.io/statsExpressions/}{package}.

The only exception is the \code{ggcorrmat()} function. But, if a data frame is
what you want, you shouldn't be using \code{ggcorrmat()} anyway. You can use
\code{correlation::correlation()} function which provides tidy data frames by
default.
}
\examples{
\dontshow{if (requireNamespace("PMCMRplus", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
set.seed(123)
library(ggstatsplot)

# non-grouped function -----------------------------

# in case of group comparisons
p <- ggbetweenstats(mtcars, cyl, mpg)
extract_stats(p)

# the exact details depend on the function
extract_stats(ggbarstats(mtcars, cyl, am))

# grouped function -----------------------------
p <- grouped_ggbarstats(
  Titanic_full,
  x = Survived,
  y = Sex,
  grouping.var = Age
)

extract_stats(p[[1L]])
extract_stats(p[[2L]])
}
\dontshow{\}) # examplesIf}
}
