% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbetweenstats.R
\name{ggbetweenstats}
\alias{ggbetweenstats}
\title{violin plots for group or condition comparisons}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{x}{the grouping variable}

\item{y}{the response - a vector of length the number of rows of \code{x}}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{type}{type of statistics expected ("parametric" or "nonparametric" or "robust")}

\item{effsize.type}{type of effect size needed for \emph{parametric} tests ("biased" (Cohen's d, partial eta-squared) or
"unbiased" (Hedge's g, omega-squared))}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{k}{number of decimal places expected for results}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal}

\item{nboot}{number of bootstrap samples}

\item{outlier.color}{default aesthetics for outliers}

\item{outlier.tagging}{whether outliers should be tagged}

\item{outlier.label}{label to put on the outliers that have been tagged; if data argument is missing}

\item{mean.plotting}{whether mean is to be highlighted and its value to be displayed
this will show y variable values for outliers}

\item{mean.color}{color for the data point corresponding to mean}
}
\description{
Violin plots for between-subjects designs with statistical details included in the plot as a subtitle
}
\examples{
# the most basic and minimalistic way of entering arguments
library(datasets)
ggstatsplot::ggbetweenstats(data = iris, x = Species, y = Sepal.Length)
# or
ggstatsplot::ggbetweenstats(x = iris$Species, y = iris$Sepal.Length)

# more detailed function call
ggstatsplot::ggbetweenstats(data = mtcars, x = cyl, y = mpg,
outlier.tagging = TRUE, outlier.label = disp, mean.plotting = TRUE)

}
\author{
Indrajeet Patil
}
