% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbetweenstats.R
\name{ggbetweenstats}
\alias{ggbetweenstats}
\title{violin plots for group or condition comparisons}
\usage{
ggbetweenstats(data = NULL, x, y, test = NULL, type = NULL,
  effsizetype = NULL, xlab = NULL, ylab = NULL, caption = NULL,
  title = NULL, k = 3, var.equal = FALSE, nboot = 1000,
  outlier.tagging = NULL, outlier.label = NULL, outlier.colour = "black",
  mean.plotting = FALSE)
}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{x}{the grouping variable}

\item{y}{the response - a vector of length the number of rows of \code{x}}

\item{test}{statistical test to be run and displayed as subtitle ("t-test" or "anova")}

\item{type}{type of statistics expected ("parametric" or "robust")}

\item{effsizetype}{type of effect size needed for parametric tests ("biased" (Cohen's d, partial eta-squared) or
"unbiased" (Hedge's g, omega-squared))}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{caption}{caption for the plot}

\item{title}{title for the plot}

\item{k}{number of decimal places expected for results}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal}

\item{nboot}{number of bootstrap samples}

\item{outlier.tagging}{whether outliers should be tagged}

\item{outlier.label}{label to put on the outliers that have been tagged; if data argument is missing}

\item{outlier.colour}{default aesthetics for outliers}

\item{mean.plotting}{whether mean is to be highlighted and its value to be displayed
this will show y variable values for outliers}
}
\author{
Indrajeet Patil
}
