% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcoefstats.R
\name{ggcoefstats_label_maker}
\alias{ggcoefstats_label_maker}
\title{Create labels with statistical details for \code{ggcoefstats}.}
\usage{
ggcoefstats_label_maker(x, tidy_df, glance_df, k = 3, effsize = "eta",
  partial = TRUE)
}
\arguments{
\item{x}{A model object to be tidied with \code{broom::tidy}.}

\item{tidy_df}{Tidy dataframe from \code{broom::tidy}.}

\item{glance_df}{Glance model summary dataframe from \code{broom::glance}.}

\item{k}{Number of decimal places expected for results displayed in labels.}

\item{effsize}{Character describing the effect size to be displayed: \code{"eta"}
(default) or \code{"omega"}. This argument is relevant
only for models objects of class \code{aov}, \code{anova}, and \code{aovlist}.}

\item{partial}{Logical that decides if partial eta-squared or omega-squared
are returned (Default: \code{TRUE}). If \code{FALSE}, eta-squared or omega-squared
will be returned. Valid only for objects of class \code{aov}, \code{anova}, or
\code{aovlist}.}
}
\description{
Create labels with statistical details for \code{ggcoefstats}.
}
\keyword{internal}
