% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{ggcorrmat_matrix_message}
\alias{ggcorrmat_matrix_message}
\title{Message to display when adjusted p-values are displayed in correlation
matrix.}
\usage{
ggcorrmat_matrix_message()
}
\description{
Message to display when adjusted p-values are displayed in correlation
matrix.
}
\seealso{
Other helper_messages: \code{\link{bartlett_message}},
  \code{\link{grouped_message}},
  \code{\link{normality_message}},
  \code{\link{pairwise_p}}, \code{\link{palette_message}}
}
\author{
Indrajeet Patil
}
\concept{helper_messages}
\keyword{internal}
