% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdotplotstats.R
\name{ggdotplotstats}
\alias{ggdotplotstats}
\alias{ggdotchartstats}
\title{Dot plot/chart for labeled numeric data.}
\usage{
ggdotplotstats(data, x, y, summarize = TRUE, xlab = NULL,
  ylab = NULL, title = NULL, subtitle = NULL, caption = NULL,
  ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)
}
\arguments{
\item{data}{Dataframe from which variables specified are preferentially to be
taken. This argument is optional.}

\item{x}{A numeric variable.}

\item{y}{Label or grouping variable.}

\item{summarize}{Logical that decides whether \code{x} is to be summarized by a
grouping variable \code{y} (Default: \code{TRUE}).}

\item{xlab}{Label for \code{x} axis variable.}

\item{ylab}{Label for \code{y} axis variable.}

\item{title}{The text for the plot title.}

\item{subtitle}{The text for the plot subtitle \emph{if} you don't want results
from one sample test to be displayed (i.e. \code{results.subtitle} should be set
to \code{FALSE}).}

\item{caption}{The text for the plot caption.}

\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_bw()}. Any of the \code{ggplot2} themes, or themes from
extension packages are allowed (e.g., \code{ggthemes::theme_economist()},
\code{hrbrthemes::theme_ipsum_ps()}, \code{ggthemes::theme_fivethirtyeight()}, etc.).}

\item{ggstatsplot.layer}{Logical that decides whether \code{theme_ggstatsplot}
theme elements are to be displayed along with the selected \code{ggtheme}
(Default: \code{TRUE}).}
}
\description{
A dot chart with statistical details from one-sample test
included in the plot as a subtitle.
}
\examples{
ggdotplotstats(
  data = ggplot2::mpg,
  x = cty,
  y = manufacturer
)
}
\author{
Indrajeet Patil
}
