% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglmstats.R
\name{gglmstats}
\alias{gglmstats}
\title{Regression coefficients for fitted linear models with the model
summary as a caption.}
\usage{
gglmstats(data, formula, k = 3, dot.color = "blue", xlab = "estimate",
  ylab = "term", title = NULL, subtitle = NULL, stats.labels = TRUE)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{formula}{An object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. The details
of model specification are given under 'Details'.}

\item{k}{Number of decimal places expected for results.}

\item{dot.color}{Character describing color for the dot (Default: \code{"blue"}).}

\item{xlab}{Label for \code{x} axis variable (Default: \code{"estimate"}).}

\item{ylab}{Label for \code{y} axis variable (Default: \code{"term"}).}

\item{title}{The text for the plot title.}

\item{subtitle}{The text for the plot subtitle.}

\item{stats.labels}{Logical. Decides whether the statistic and p-values for
each coefficient are to be attached to each dot as a text label using
\code{ggrepel}.}
}
\value{
Plot with the regression coefficients' point estimates as dots with
confidence interval whiskers.
}
\description{
Regression coefficients for fitted linear models with the model
summary as a caption.
}
\examples{

set.seed(123)
gglmstats(formula = mpg ~ cyl * am, data = mtcars)

}
\author{
Indrajeet Patil
}
