% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpiestats.R
\name{ggpiestats}
\alias{ggpiestats}
\title{pie charts with statistical tests}
\usage{
ggpiestats(data, main, condition = NULL, labels = NULL, stat_title = NULL,
  title = NULL, caption = NULL, legend_title = NULL, k = 3)
}
\arguments{
\item{data}{the data as a data frame}

\item{main}{a string naming the variable to use as the rows in the contingency table}

\item{condition}{a string naming the variable to use as the columns in the contingency table}

\item{labels}{a character vector of same length as (data, main) or (data, condition, main) if facet wrapping}

\item{stat_title}{title for the effect being investigated with the chi-square test}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{k}{number of decimal places expected for results}

\item{type}{type of statistics expected ("parametric" or "robust")}
}
\author{
Indrajeet Patil
}
