% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpiestats.R
\name{ggpiestats}
\alias{ggpiestats}
\title{pie charts with statistical tests}
\arguments{
\item{data}{The data as a data frame.}

\item{main}{A string naming the variable to use as the rows in the
contingency table.}

\item{condition}{A string naming the variable to use as the columns in the
contingency table.}

\item{factor.levels}{A character vector with labels for factor levels of
\code{main} variable.}

\item{stat.title}{Title for the effect being investigated with the chi-square
test.}

\item{title}{The text for the plot title.}

\item{caption}{The text for the plot caption.}

\item{k}{Number of decimal places expected for results.}

\item{legend.title}{Title of legend.}

\item{facet.wrap.name}{The text for the facet_wrap variable label.}

\item{messages}{Decides whether messages references, notes, and warnings are
to be displayed (Default: \code{TRUE}).}
}
\description{
Pie charts for categorical data with statistical details
included in the plot as a subtitle.
}
\examples{
library(datasets)
ggstatsplot::ggpiestats(data = mtcars, main = am, condition = cyl, factor.levels = c("zero", "one"),
stat.title = "interaction", title = "Dataset: mtcars", caption = "This is a demo",
legend.title = "transmission", facet.wrap.name = "No. of cylinders")

}
\author{
Indrajeet Patil
}
