% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpiestats.R
\name{ggpiestats}
\alias{ggpiestats}
\title{pie charts with statistical tests}
\usage{
ggpiestats(data = NULL, main, condition = NULL, labels = NULL,
  stat.title = NULL, title = NULL, caption = NULL, legend.title = NULL,
  k = 3)
}
\arguments{
\item{data}{the data as a data frame}

\item{main}{a string naming the variable to use as the rows in the contingency table}

\item{condition}{a string naming the variable to use as the columns in the contingency table}

\item{stat.title}{title for the effect being investigated with the chi-square test}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{legend.title}{title for the legend}

\item{k}{number of decimal places expected for results}
}
\description{
Pie charts for categorical data with statistical details included in the plot as a subtitle
}
\examples{
library(datasets)
ggpiestats(data = iris, main = Species)
# or
ggpiestats(main = iris$Species)
# with condition variable
ggpiestats(data = mtcars, main = am, condition = cyl)

}
\author{
Indrajeet Patil
}
