% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatterstats.R
\name{ggscatterstats}
\alias{ggscatterstats}
\title{scatterplot with ggMarginals}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{x}{a vector containing the explanatory variable}

\item{y}{the response - a vector of length the number of rows of \code{x}}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{marginal}{decides whether \code{ggExtra::ggMarginal()} plots will be displayed; the default is \code{TRUE}}

\item{marginal.type}{type of marginal distribution to be plotted on the axes
("histogram", "boxplot", "density", "violin")}

\item{xfill}{colour fill for x axis distibution (default: "orange")}

\item{yfill}{colour fill for y axis distribution (default: "green")}

\item{test}{statistical test to be run and displayed as subtitle ("pearson", "spearman", "robust")}

\item{results.subtitle}{whether the results of statistical tests are to be displayed as subtitle}

\item{intercept}{decides whether "mean" or "median" or no intercept lines (\code{NULL}) are to be plotted}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{maxit}{maximum number of iterations for robust linear regression}

\item{k}{number of decimal places expected for results}

\item{width.jitter}{degree of jitter in x direction. Defaults to 40\% of the resolution of the data.}

\item{height.jitter}{degree of jitter in y direction. Defaults to 40\% of the resolution of the data}
}
\description{
Scatterplots from `ggplot2`` combined with add marginal histograms/boxplots/density plots with
statistical details added as a subtitle
}
\note{
If you want to use \code{marginal.type = "violin"}, you will have to use development version of \code{ggExtra}
}
\examples{
# the most basic and minimalistic way of entering arguments
library(datasets)
ggscatterstats(data = iris, x = Petal.Length, y = Sepal.Length)
# or
ggscatterstats(x = iris$Petal.Length, y = iris$Sepal.Length)

# more detailed call
ggscatterstats(x = iris$Petal.Length, y = iris$Sepal.Length,
intercept = 'median', test = 'robust', marginal.type = 'density')

}
\author{
Indrajeet Patil
}
