% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatterstats.R
\name{ggscatterstats}
\alias{ggscatterstats}
\title{scatterplot with ggMarginals}
\usage{
ggscatterstats(data = NULL, x, y, xlab = NULL, ylab = NULL,
  marginal = NULL, marginaltype = NULL, xfill = NULL, yfill = NULL,
  intercept = NULL, test = NULL, title = NULL, caption = NULL,
  maxit = 1000, k = 3)
}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{x}{a vector containing the explanatory variable}

\item{y}{the response - a vector of length the number of rows of \code{x}}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{marginal}{decides whether \code{ggExtra::ggMarginal()} plots will be displayed; the default is \code{TRUE}}

\item{marginaltype}{type of marginal distribution to be plotted on the axes (NULL, "histogram", "boxplot", "density")}

\item{xfill}{color fill for x axis distibution}

\item{yfill}{color fill for y axis distribution}

\item{intercept}{decides whether "mean" or "median" or no intercept lines (\code{NULL}) are to be plotted}

\item{test}{statistical test to be run and displayed as subtitle ("pearson", "spearman", "robust")}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{maxit}{maximum number of iterations for robust linear regression}

\item{k}{number of decimal places expected for results}
}
\description{
Scatterplots from `ggplot2`` combined with add marginal histograms/boxplots/density plots with
statistical details added as a subtitle
}
\author{
Indrajeet Patil
}
