% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscatterstats.R
\name{ggscatterstats}
\alias{ggscatterstats}
\title{scatterplot with ggMarginals}
\usage{
ggscatterstats(data = NULL, x, y, xlab = NULL, ylab = NULL,
  marginal = NULL, marginaltype = NULL, xfill = "orange",
  yfill = "green", intercept = NULL, test = NULL,
  results.subtitle = NULL, title = NULL, caption = NULL, maxit = 1000,
  k = 3, jitter.width = NULL, jitter.height = 0.2, dodge.width = 0.75)
}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{x}{a vector containing the explanatory variable}

\item{y}{the response - a vector of length the number of rows of \code{x}}

\item{xlab}{label for x axis variable}

\item{ylab}{label for y axis variable}

\item{marginal}{decides whether \code{ggExtra::ggMarginal()} plots will be displayed; the default is \code{TRUE}}

\item{marginaltype}{type of marginal distribution to be plotted on the axes ("histogram", "boxplot", "density", "violin")}

\item{xfill}{colour fill for x axis distibution (default: "orange")}

\item{yfill}{colour fill for y axis distribution (default: "green")}

\item{intercept}{decides whether "mean" or "median" or no intercept lines (\code{NULL}) are to be plotted}

\item{test}{statistical test to be run and displayed as subtitle ("pearson", "spearman", "robust")}

\item{results.subtitle}{whether the results of statistical tests are to be displayed as subtitle}

\item{title}{title for the plot}

\item{caption}{caption for the plot}

\item{maxit}{maximum number of iterations for robust linear regression}

\item{k}{number of decimal places expected for results}

\item{jitter.width}{degree of jitter in x direction. Defaults to 0.40 of the resolution of the data}

\item{jitter.height}{degree of jitter in y direction. Defaults to 0}

\item{dodge.width}{the amount to dodge in the x direction. Defaults to 0.75, the default \code{position_dodge()} width}
}
\description{
Scatterplots from `ggplot2`` combined with add marginal histograms/boxplots/density plots with
statistical details added as a subtitle
}
\examples{
# the most basic and minimalistic way of entering arguments
library(datasets)
ggscatterstats(data = iris, x = Petal.Length, y = Sepal.Length)
# or
ggscatterstats(x = iris$Petal.Length, y = iris$Sepal.Length)

# more detailed call
ggscatterstats(x = iris$Petal.Length, y = iris$Sepal.Length,
intercept = 'median', test = 'robust', marginaltype = 'density')

}
\author{
Indrajeet Patil
}
