% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats_graphics.R
\name{ggsignif_position_calculator}
\alias{ggsignif_position_calculator}
\title{Calculating \code{y} coordinates for the \code{ggsignif} comparison bars.}
\usage{
ggsignif_position_calculator(x, y)
}
\arguments{
\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}
}
\description{
Calculating \code{y} coordinates for the \code{ggsignif} comparison bars.
}
\keyword{internal}
