% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstatsplot-package.R
\docType{package}
\name{ggstatsplot-package}
\alias{ggstatsplot-package}
\title{\code{ggstatsplot}}
\description{
Collection of functions to enhance \code{ggplot2} plots with results from
statistical tests.
}
\details{
The main functions are- \itemize{ \item
\code{\link[ggstatsplot]{ggbetweenstats}} function to produce
information-rich comparison plot between different groups or conditions
with \code{ggplot2} and details from the statistical tests in the subtitle \item
\code{\link[ggstatsplot]{ggscatterstats}} function to produce \code{ggplot2}
scatterplots along with a marginal histograms/boxplots/density plots from
\code{ggExtra} and details from the statistical tests in the subtitle \item
\code{\link[ggstatsplot]{ggpiestats}} function to produce pie chart with
details from the statistical tests in the subtitle \item
\code{\link[ggstatsplot]{gghistostats}} function to produce histogram for a
single variable with results from one sample test displayed in the subtitle
\item \code{\link[ggstatsplot]{ggcorrmat}} function to visualize
correlation matrix \item \code{\link[ggstatsplot]{combine_plots}} helper
function to combine multiple \code{ggstatsplot} plots using
\code{cowplot::plot_grid()} with a combination of title, caption, and annotation
label \item \code{\link[ggcoefstats]{ggcoefstats}} function to visualize
results from regression analyses \item
\code{\link[ggstatsplot]{theme_ggstatsplot}} default theme used for this
package \item \code{\link[ggstatsplot]{specify_decimal_p}} helper function
to format results for pretty printing }

For more documentation, see
\href{https://github.com/IndrajeetPatil/ggstatsplot}{README} on GitHub.
Vignette preparation in progress.
}
