% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_gghistostats.R
\name{grouped_gghistostats}
\alias{grouped_gghistostats}
\title{Grouped histograms for distribution of a numeric variable}
\arguments{
\item{grouping.vars}{List of grouping variables.}

\item{...}{Arguments passed on to \code{combine_plots}
\describe{
  \item{title.text}{String or plotmath expression to be drawn as title for the \emph{combined plot}.}
  \item{title.colour}{Text color for title.}
  \item{title.size}{Point size of title text.}
  \item{title.vjust}{Vertical justification for title. Default = 0.5 (centered on y).
0 = baseline at y, 1 = ascender at y.}
  \item{title.hjust}{Horizontal justification for title. Default = 0.5 (centered on x).
0 = flush-left at x, 1 = flush-right.}
  \item{title.fontface}{The font face ("plain", "bold", etc.) for title.}
  \item{caption.text}{String or plotmath expression to be drawn as the caption for the \emph{combined plot}.}
  \item{caption.colour}{Text color for caption.}
  \item{caption.size}{Point size of title text.}
  \item{caption.vjust}{Vertical justification for caption. Default = 0.5 (centered on y).
0 = baseline at y, 1 = ascender at y.}
  \item{caption.hjust}{Horizontal justification for caption. Default = 0.5 (centered on x).
0 = flush-left at x, 1 = flush-right.}
  \item{caption.fontface}{The font face ("plain", "bold", etc.) for caption.}
  \item{sub.text}{The label with which the \emph{combined plot} should be annotated. Can be a plotmath expression.}
  \item{sub.colour}{Text color for annotation label.}
  \item{sub.size}{Point size of annotation text.}
  \item{sub.x}{The x position of annotation label.}
  \item{sub.y}{The y position of annotation label.}
  \item{sub.hjust}{Horizontal justification for annotation label.}
  \item{sub.vjust}{Vertical justification for annotation label.}
  \item{sub.vpadding}{Vertical padding. The total vertical space added to the label, given in grid units.
By default, this is added equally above and below the label. However, by changing the y and vjust parameters, this can be changed.}
  \item{sub.fontface}{The font face ("plain", "bold", etc.) for the annotation label.}
  \item{sub.angle}{Angle at which annotation label is to be drawn.}
  \item{sub.lineheight}{Line height of annotation label.}
  \item{title.caption.rel.heights}{Numerical vector of relative columns heights while combining (title, plot, caption).}
  \item{title.rel.heights}{Numerical vector of relative columns heights while combining (title, plot).}
  \item{caption.rel.heights}{Numerical vector of relative columns heights while combining (plot, caption).}
}}
}
\description{
Helper function for \code{ggstatsplot::gghistostats} to apply this
function across multiple levels of a given factor and combining the
resulting plots using \code{ggstatsplot::combine_plots}.
}
\examples{

library(gapminder)

ggstatsplot::grouped_gghistostats(
data = gapminder::gapminder,
x = lifeExp,
xlab = "Life expectancy",
test.value = 50,
grouping.vars = continent,
nrow = 3
)

}
\author{
Indrajeet Patil
}
