% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grouped_proptest}
\alias{grouped_proptest}
\title{Function to run proportion test on grouped data.}
\arguments{
\item{data}{Dataframe from which variables are to be drawn.}

\item{grouping.vars}{List of grouping variables}

\item{measure}{A variable for which proportion test needs to be carried out
for each combination of levels of factors entered in \code{grouping.vars}.}
}
\value{
Dataframe with percentages and statistical details from a proportion
test.
}
\description{
Function to run proportion test on grouped data.
}
\author{
Indrajeet Patil
}
\keyword{internal}
