% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_w_ci.R
\name{kendall_w_ci}
\alias{kendall_w_ci}
\title{Computing confidence intervals for the Kendall's coefficient of
concordance (aka Kendall's \emph{W}).}
\usage{
kendall_w_ci(data, x, y, id.variable, conf.level = 0.95, ...)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{id.variable}{Variable identifying a unique observation.}

\item{conf.level}{Scalar between 0 and 1. If unspecified, the defaults return
\code{95\%} lower and upper confidence intervals (\code{0.95}).}

\item{...}{Additional arguments.}
}
\description{
Computing confidence intervals for the Kendall's coefficient of
concordance (aka Kendall's \emph{W}).
}
\examples{

set.seed(123)

ggstatsplot:::kendall_w_ci(
  data = iris_long,
  x = condition,
  y = value,
  id.variable = id
)
}
\author{
Chuck Powell, Indrajeet Patil
}
\keyword{internal}
