% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcoefstats_graphics.R
\name{lm_effsize_standardizer}
\alias{lm_effsize_standardizer}
\title{Tidying anova-like objects}
\usage{
lm_effsize_standardizer(object, effsize = "eta", conf.level = 0.95, ...)
}
\arguments{
\item{object}{The object of the following class: \code{aov}, \code{aovlist}, \code{anova},
\code{Gam}, \code{manova}, \code{maov}.}

\item{effsize}{Character describing the effect size to be displayed: \code{"eta"}
(default) or \code{"omega"}.}

\item{conf.level}{Numeric specifying Level of confidence for the confidence
interval (Default: \code{0.95}).}

\item{...}{Ignored.}
}
\description{
Tidying anova-like objects
}
\examples{
# for reproducibility
set.seed(123)

# model
mod <-
  stats::aov(
    formula = mpg ~ wt + qsec + Error(disp / am),
    data = mtcars
  )

# dataframe with effect size and confidence intervals
ggstatsplot:::lm_effsize_standardizer(mod)
}
\keyword{internal}
