% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggbetweenstats.R
\name{mean_labeller}
\alias{mean_labeller}
\title{@title Create a dataframe with mean per group and a formatted label for
display in \code{ggbetweenstats} plot.}
\usage{
mean_labeller(data, x, y, mean.ci = FALSE, k = 3)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{mean.ci}{Logical that decides whether 95% confidence interval for mean
is to be displayed (Default: \code{FALSE}).}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2}).}
}
\description{
@title Create a dataframe with mean per group and a formatted label for
display in \code{ggbetweenstats} plot.
}
\examples{
\donttest{
ggstatsplot:::mean_labeller(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  mean.ci = TRUE,
  k = 3
)
}

}
\keyword{internal}
