% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{proptest_message}
\alias{proptest_message}
\title{Message about results from a single-sample proportion test.}
\usage{
proptest_message(main, condition)
}
\arguments{
\item{main, condition}{Character specifying names of variables used for
contingency table analyses.}
}
\description{
Message about results from a single-sample proportion test.
}
\examples{
\donttest{
ggstatsplot:::proptest_message(main = "am", condition = "cyl")
}

}
\author{
Indrajeet Patil
}
\keyword{internal}
