% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%<-\%}
\alias{tibble}
\alias{pairwise_comparisons}
\alias{as_tibble}
\alias{exec}
\alias{!!}
\alias{!!!}
\alias{\%|\%}
\alias{\%||\%}
\alias{:=}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:dyn-dots]{:=}}, \code{\link[rlang:injection-operator]{!!}}, \code{\link[rlang:splice-operator]{!!!}}, \code{\link[rlang:op-na-default]{\%|\%}}, \code{\link[rlang:op-null-default]{\%||\%}}, \code{\link[rlang]{exec}}}

  \item{statsExpressions}{\code{\link[statsExpressions:reexports]{\%<-\%}}, \code{\link[statsExpressions:reexports]{\%<>\%}}, \code{\link[statsExpressions:reexports]{\%>\%}}, \code{\link[statsExpressions:reexports]{\%$\%}}, \code{\link[statsExpressions:reexports]{as_tibble}}, \code{\link[statsExpressions]{pairwise_comparisons}}, \code{\link[statsExpressions:reexports]{tibble}}}
}}

