% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{specify_decimal_p}
\alias{signif_column}
\alias{set_cwd}
\alias{pairwise_p}
\alias{pairwise_comparisons}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%<-\%}
\alias{\%||\%}
\alias{\%|\%}
\alias{tibble}
\alias{as_tibble}
\alias{bartlett_message}
\alias{normality_message}
\alias{tidy}
\alias{glance}
\alias{augment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broomExtra}{\code{\link[broomExtra]{augment}}, \code{\link[broomExtra]{glance}}, \code{\link[broomExtra]{tidy}}}

  \item{ipmisc}{\code{\link[ipmisc]{\%$\%}}, \code{\link[ipmisc]{\%<-\%}}, \code{\link[ipmisc]{\%<>\%}}, \code{\link[ipmisc]{\%>\%}}, \code{\link[ipmisc]{as_tibble}}, \code{\link[ipmisc]{bartlett_message}}, \code{\link[ipmisc]{normality_message}}, \code{\link[ipmisc]{set_cwd}}, \code{\link[ipmisc]{signif_column}}, \code{\link[ipmisc]{specify_decimal_p}}, \code{\link[ipmisc]{tibble}}}

  \item{pairwiseComparisons}{\code{\link[pairwiseComparisons]{pairwise_comparisons}}, \code{\link[pairwiseComparisons]{pairwise_p}}}

  \item{rlang}{\code{\link[rlang]{\%|\%}}, \code{\link[rlang]{\%||\%}}}
}}

