% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{specify_decimal_p}
\alias{signif_column}
\alias{lm_effsize_ci}
\alias{set_cwd}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%||\%}
\alias{\%|\%}
\alias{tibble}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broomExtra}{\code{\link[broomExtra]{\%$\%}}, \code{\link[broomExtra]{\%<>\%}}, \code{\link[broomExtra]{\%>\%}}}

  \item{groupedstats}{\code{\link[groupedstats]{lm_effsize_ci}}, \code{\link[groupedstats]{set_cwd}}, \code{\link[groupedstats]{signif_column}}, \code{\link[groupedstats]{specify_decimal_p}}}

  \item{rlang}{\code{\link[rlang]{\%|\%}}, \code{\link[rlang]{\%||\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}
}}

