% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal_p.R
\name{specify_decimal_p}
\alias{specify_decimal_p}
\title{custom function for getting specified number of decimal places in results for p-value}
\usage{
specify_decimal_p(x, k = NULL, p.value = FALSE)
}
\arguments{
\item{x}{A numeric variable.}

\item{k}{Number of digits after decimal point (should be an integer).}

\item{p.value}{Decides whether the number is a p-value (Dafault: \code{FALSE}).}
}
\value{
formatted p-values from statistical tests
}
\description{
Function to format an R object for pretty printing with a specified number of decimal places. The
function also allows highly significant p-values to be denoted as "p < 0.001" rather than "p = 0.000".
}
\author{
Indrajeet Patil
}
