% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggcoefstats.R
\name{tfz_labeller}
\alias{tfz_labeller}
\title{Create text labels}
\usage{
tfz_labeller(tidy_df, glance_df = NULL, statistic, effsize = "eta",
  partial = TRUE, k)
}
\arguments{
\item{tidy_df}{Tidy dataframe from \code{broom::tidy}.}

\item{glance_df}{Glance model summary dataframe from \code{broom::glance}.}

\item{statistic}{Which statistic is to be displayed (either \code{"t"} or \code{"f"}or
\code{"z"}).}

\item{effsize}{Character describing the effect size to be displayed: \code{"eta"}
(default) or \code{"omega"}.}

\item{partial}{Logical that decides if partial eta-squared or omega-squared
are returned (Default: \code{TRUE}). If \code{FALSE}, eta-squared or omega-squared
will be returned. Valid only for objects of class \code{lm}, \code{aov}, \code{anova}, or
\code{aovlist}.}

\item{k}{Number of decimal places expected for results displayed in labels.}
}
\description{
Create text labels
}
\keyword{internal}
