% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_graphics.R
\name{theme_corrmat}
\alias{theme_corrmat}
\title{Default theme used for correlation matrix}
\usage{
theme_corrmat()
}
\value{
A \code{ggplot2} object with the \code{theme_mprl} theme.
}
\description{
Default theme used for correlation matrix
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::theme_corrmat}. Note that it is \code{:::} and not \code{::}.
}
\author{
Indrajeet Patil
#'
}
\keyword{internal}
