% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mprl.R
\name{theme_ggstatsplot}
\alias{theme_ggstatsplot}
\title{Default theme used in all \code{ggstatsplot} package plots}
\usage{
theme_ggstatsplot(ggtheme = ggplot2::theme_bw())
}
\arguments{
\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_bw()}. Allowed values are the official \code{ggplot2} themes,
including \code{theme_bw()}, \code{theme_minimal()}, \code{theme_classic()},
\code{theme_void()}, etc.}
}
\description{
Common theme used across all plots generated in \code{ggstatsplot}
and \emph{assumed} by the author to be aesthetically pleasing to the
user/reader.
}
\note{
This is just another name for the function \code{theme_mprl()}.
}
\references{
\url{https://indrajeetpatil.github.io/ggstatsplot/articles/theme_ggstatsplot.html}
}
\author{
Indrajeet Patil
}
