% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_theme_ggstatsplot.R
\name{theme_ggstatsplot}
\alias{theme_ggstatsplot}
\title{Default theme used in all \code{ggstatsplot} package plots}
\usage{
theme_ggstatsplot()
}
\value{
A \code{ggplot2} object with the \code{theme_ggstatsplot} theme overlaid.
}
\description{
Common theme used across all plots generated in \code{ggstatsplot} and \emph{assumed}
by the author to be aesthetically pleasing to the user/reader.
}
\details{
For more details, see:
\url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/theme_ggstatsplot.html}
}
