% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_ggstatsplot}
\alias{theme_ggstatsplot}
\title{Default theme used in \code{{ggstatsplot}}}
\usage{
theme_ggstatsplot()
}
\value{
A \code{ggplot} object with the \code{theme_ggstatsplot} theme overlaid.
}
\description{
Common theme used across all plots generated in \code{{ggstatsplot}} and \emph{assumed}
by the author to be aesthetically pleasing to the user/reader. The theme is a
wrapper around \code{theme_bw()}.
}
\examples{
library(ggplot2)
library(ggstatsplot)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_ggstatsplot()
}
