% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggstatsplot.R
\name{theme_ggstatsplot}
\alias{theme_ggstatsplot}
\alias{theme_mprl}
\title{Default theme used in all \code{ggstatsplot} package plots}
\usage{
theme_ggstatsplot(ggtheme = ggplot2::theme_bw(),
  ggstatsplot.layer = TRUE)

theme_mprl(ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)
}
\arguments{
\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_bw()}. Any of the \code{ggplot2} themes, or themes from
extension packages are allowed (e.g., \code{ggthemes::theme_fivethirtyeight()},
\code{hrbrthemes::theme_ipsum_ps()}, etc.).}

\item{ggstatsplot.layer}{Logical that decides whether \code{theme_ggstatsplot}
theme elements are to be displayed along with the selected \code{ggtheme}
(Default: \code{TRUE}).}
}
\value{
A \code{ggplot2} object with the \code{theme_ggstatsplot} theme.
}
\description{
Common theme used across all plots generated in \code{ggstatsplot}
and \emph{assumed} by the author to be aesthetically pleasing to the
user/reader.
}
\references{
\url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/theme_ggstatsplot.html}
}
\author{
Indrajeet Patil
}
