% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mprl.R
\name{theme_mprl}
\alias{theme_mprl}
\alias{theme_ggstatsplot}
\title{Default theme used in all \code{ggstatsplot} package plots}
\usage{
theme_mprl()
}
\value{
A \code{ggplot2} object with the \code{theme_mprl} theme.
}
\description{
Common theme used across all plots generated in \code{ggstatsplot}
and \emph{assumed} by the author to be aesthetically pleasing to the
user/reader.
}
\author{
Indrajeet Patil
}
