% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggstatsplot.R
\name{theme_pie}
\alias{theme_pie}
\title{Default theme used for pie chart}
\usage{
theme_pie(ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)
}
\arguments{
\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_bw()}. Any of the \code{ggplot2} themes, or themes from
extension packages are allowed (e.g., \code{ggthemes::theme_economist()},
\code{hrbrthemes::theme_ipsum_ps()}, \code{ggthemes::theme_fivethirtyeight()}, etc.).}

\item{ggstatsplot.layer}{Logical that decides whether \code{theme_ggstatsplot}
theme elements are to be displayed along with the selected \code{ggtheme}
(Default: \code{TRUE}).}
}
\value{
A \code{ggplot2} object with the \code{theme_ggstatsplot} theme.
}
\description{
Default theme used for pie chart
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::theme_pie}. Note that it is \code{:::} and not \code{::}.
}
\author{
Indrajeet Patil
}
\keyword{internal}
