% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_theme_ggstatsplot.R
\name{theme_pie}
\alias{theme_pie}
\title{Default theme used for pie chart}
\usage{
theme_pie(ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)
}
\arguments{
\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_bw()}. Any of the \code{ggplot2} themes, or themes from
extension packages are allowed (e.g., \code{ggthemes::theme_fivethirtyeight()},
\code{hrbrthemes::theme_ipsum_ps()}, etc.).}

\item{ggstatsplot.layer}{Logical that decides whether \code{theme_ggstatsplot}
theme elements are to be displayed along with the selected \code{ggtheme}
(Default: \code{TRUE}). \code{theme_ggstatsplot} is an opinionated theme layer that
override some aspects of the selected \code{ggtheme}.}
}
\value{
A \code{ggplot2} object with the \code{theme_ggstatsplot} theme.
}
\description{
Default theme used for pie chart
}
