% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_graphics.R
\name{theme_pie}
\alias{theme_pie}
\title{Default theme used for pie chart}
\usage{
theme_pie(ggtheme = ggplot2::theme_grey())
}
\arguments{
\item{ggtheme}{A function, \code{ggplot2} theme name. Default value is
\code{ggplot2::theme_grey()}. Allowed values are the official \code{ggplot2} themes,
including \code{theme_bw()}, \code{theme_minimal()}, \code{theme_classic()},
\code{theme_void()}, etc.}
}
\value{
A \code{ggplot2} object with the \code{theme_mprl} theme.
}
\description{
Default theme used for pie chart
}
\note{
This is a helper function used internally in the package and not
exported. In case you want to use it, you can do so by
\code{ggstatsplot:::theme_pie}. Note that it is \code{:::} and not \code{::}.
}
\author{
Indrajeet Patil
#'
}
\keyword{internal}
